[NuGet] Fix requests bypasing the native HTTP handler
Restoring or installing a NuGet package into a project that uses
PackageReferences would bypass the native HTTP handler. This was due
to NuGet's CachingSourceProvider creating its own set of
INuGetResourceProviders when it creates a set of new SourceRepository
instances. This resulted in the default HttpSourceResourceProvider
being used when uses Mono's HttpClientHandler. This would happen
when the .nupkg was downloaded.

NuGet 5.4 allows a custom NuGetResourceProvider factory to be
configured instead of only when creating instances of SourceRepository.
This is now used by the Package Management addin so all HTTP requests
use the native HTTP handler.

Fixes VSTS #918815 - Fixes NuGet requests bypassing native Http Handler