[NuGet] Show license metadata in License Acceptance dialog
If the NuGet package has associated license metadata then show this
in the dialog. License metadata can be an expression or a file.
A license file will now be displayed in a dialog if the View License
link is clicked. Warning icons and associated message will also be
displayed for deprecated licenses.

There are problems here. Could not get the text to wrap so ended up
making the dialog resizable and always showing the scrollbars.

Hyperlinks in license expressions do not open the url in a browser
since the Xwt Xamarin.Mac's LabelBackend does not implement support
for this.